Made ChannelIdx uint16.
To unify channel index type: uint8 is enough for us4oems, however it wont work for adapters and probes.

+++ core/api/common/types.h
@@ -10,7 +10,7 @@ namespace arrus {
 	using uint16 = unsigned short;
     using uint8 = unsigned char;
 
-	using ChannelIdx = uint8;
+	using ChannelIdx = uint16;
 	using BitMask = std::vector<bool>;
 
 }
